refactor(web): renomeia o componte
